+++ src/main/java/lotto/view/InputView.java
@@ -1,17 +1,23 @@
 package lotto.view;
 
+import lotto.domain.PurchaseBudget;
+
+import static lotto.constants.InputMessage.DEMAND_INPUT_BUDGET;
+import static lotto.util.InputUtils.readInt;
+
 public class InputView {
 
+    public static PurchaseBudget getPurchaseBudget() {
+        try {
+            println(DEMAND_INPUT_BUDGET.getMessage());
+            return new PurchaseBudget(readInt());
+        } catch (IllegalArgumentException e) {
+            println(e.getMessage());
+            return getPurchaseBudget();
+        }
+    }
+
     public static void println(String message) {
-        print(message);
-        lineBreak();
-    }
-
-    public static void print(String message) {
-        System.out.print(message);
-    }
-
-    public static void lineBreak() {
-        System.out.println();
+        System.out.println(message);
     }
 }
