+++ src/main/java/lotto/view/OutputView.java
@@ -1,7 +1,22 @@
 package lotto.view;
 
+import lotto.constants.OutputMessage;
+import lotto.domain.PurchaseQuantity;
+
+import static lotto.constants.OutputMessage.ANNOUNCE_PURCHASE_QUANTITY_FORMAT;
+
 public class OutputView {
-    public static void println(String message) {
-        System.out.println(message);
+
+    public static void announcePurchaseQuantity(PurchaseQuantity purchaseQuantity) {
+        printObject(ANNOUNCE_PURCHASE_QUANTITY_FORMAT, purchaseQuantity);
+        lineBreak();
+    }
+
+    public static void printObject(OutputMessage outputMessage, Object object) {
+        System.out.print(String.format(outputMessage.getMessage(), object.toString()));
+    }
+
+    public static void lineBreak() {
+        System.out.println();
     }
 }
